@@ -8,11 +8,15 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3.0",
         "fideloper/proxy": "^4.0",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0"
+        "laravel/ui": "^3.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "facade/ignition": "^2.3.6.",
+        "nunomaduro/collision": "^5.0"
+
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.2",
@@ -22,9 +26,8 @@
         "fzaninotto/faker": "^1.4",
         "jason-guru/laravel-make-repository": "^0.0.3",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5",
-        "facade/ignition": "^2.0"
+        "phpunit/phpunit": "^9.0"
+
     },
     "config": {
         "optimize-autoloader": true,
@@ -38,12 +41,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
