@@ -1,6 +1,9 @@
 <?php
+namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
+use App\Models\BlogPost;
+
 
 class DatabaseSeeder extends Seeder
 {
@@ -13,6 +16,7 @@ class DatabaseSeeder extends Seeder
     {
          $this->call(UserTableSeeder::class);
          $this->call(BlogCategoriesTableSeeder::class);
-        factory(App\Models\BlogPost::class, 100)->create();
+         BlogPost::factory()->count(100)->create();
+//        factory(App\Models\BlogPost::class, 100)->create();
     }
 }
